requirements.txt
crewai==0.11.2
langchain==0.1.20
langchain-openai==0.0.5
langchain-community==0.0.38
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
numpy==2.4.1
